requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
libretranslatepy==2.1.1
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
pip==24.3.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
setuptools==65.5.0
SQLAlchemy==2.0.36
translate==3.6.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0